# Allow JAX objects to be represented by multiple buffers