--- a/jax/interpreters/pxla.py
+++ b/jax/interpreters/pxla.py
@@ -66,6 +66,15 @@ unsafe_map, map = map, safe_map
 Index = Union[int, slice, Tuple[Union[int, slice], ...]]
 
 
+def device_put(x, devices: Sequence[xb.xla_client.Device], replicate: bool=False) -> List[xb.xla_client._xla.PyLocalBuffer]:
+  """Call device_put on a sequence of devices and return a flat sequence of buffers."""
+  if replicate:
+    return list(it.chain.from_iterable(xla.device_put(x, device) for device in devices))
+  else:
+    return list(it.chain.from_iterable(xla.device_put(val, device) for val, device in safe_zip(x, devices)))
+
+
+
 # TODO(skye): make this a namedtuple. This may allow us to use ShardingSpecs in
 # performance-sensitive code, e.g. shard_args.
 class ShardingSpec:
@@ -237,9 +246,9 @@ def shard_args(devices: Sequence[xb.xla_client.Device],
 
 shard_arg_handlers: Dict[Any, Callable[[Any, Any, Any], Sequence[Any]]] = {}
 shard_arg_handlers[core.Unit] = \
-    lambda x, devices, _: [xla.device_put(core.unit, d) for d in devices]
+    lambda x, devices, _: device_put(core.unit, devices, replicate=True)
 def _shard_array(x, devices, indices):
-  return [xla.device_put(x[i], d) for (i, d) in zip(indices, devices)]
+  return device_put([x[i] for i in indices], devices)
 for _t in array_types:
   shard_arg_handlers[_t] = _shard_array
 
@@ -247,7 +256,7 @@ def _shard_device_array(x, devices, indices):
   start_indices, limit_indices, removed_dims = map(tuple, unzip3(
       _as_slice_indices(x, idx) for idx in indices))
   shards = x._multi_slice(start_indices, limit_indices, removed_dims)
-  return [xla.device_put(s, d) for s, d in zip(shards, devices)]
+  return device_put(shards, devices)
 shard_arg_handlers[xla.DeviceArray] = _shard_device_array
 
 # NOTE(skye): we could refactor to generate _multi_slice parameters directly
@@ -379,6 +388,9 @@ class ShardedDeviceArray(xla.DeviceArray):
     if indices is None:
       indices = spec_to_indices(aval.shape, sharding_spec)
     self.aval = aval
+    # TODO(jakevdp): remove this check once multi-buf has baked.
+    if not all(isinstance(b, xc.Buffer) for b in device_buffers):
+      raise TypeError(f"device_buffers should be a flat list of Buffers; got {device_buffers}")
     self.device_buffers = device_buffers
     self.sharding_spec = sharding_spec
     self.indices = indices
@@ -862,7 +874,7 @@ def replicate(val, axis_size, nrep, devices=None, backend=None):
   replicated_aval = ShapedArray((axis_size,) + aval.shape, aval.dtype)
   # TODO(skye): figure out how partitioning should work here
   sharding_spec = _pmap_sharding_spec(nrep, axis_size, 1, None, aval, True)
-  device_buffers = [xla.device_put(val, d) for d in devices]
+  device_buffers = device_put(val, devices, replicate=True)
   return ShardedDeviceArray(replicated_aval, sharding_spec, device_buffers)
 
 def _pmap_sharding_spec(nrep, axis_size, npart, parts, sharded_aval, mapped):