… tests/custom_object_test.py