Why? This will allow us to support more flexible extensions of JAX types, such as sparse matrices (which are single objects wrapping multiple buffers).

SInce this change touches large parts of the package, I plan to approach it incrementally, changing a bit at a time and ensuring tests pass:

- [x] make `xla.device_put()` return a length-1 tuple of buffers
- [x] make `xla.aval_to_xla_shape` return length-1 tuple of xla_shapes
- [x] partition flattened buffers properly for xla result handlers
- [x] generalize above to multiple buffers & add a test of a basic jax object backed by two buffers
- [ ] ~modify `unit` to return an empty tuple instead of a place-holder~ ...may be possible, but is not trivial. For example, there are multiple flavors of object that use `AbstractUnit` as their `aval`, including `Unit`, `UnitVar`, and `DropVar`. The latter in particular cannot be lowered to zero buffers, becuase it is used to ingest (and then ignore) a buffer produced by a primitive that returns multiple buffers, when that buffer happens to be unused in the jaxpr. I tried working around this by creating a custom abstract value for `DropVar`, but there remained some failures related to `UnitVar` for which it was difficult to find the root cause.
- [ ] ~possible to remove `multiple_results` from primitives?~ ...probably not; we still want some primitives to return values and some to return tuples, though we need to be careful with flattening when there are multiple results that have multiple buffers (covered by lax loop test case)