@jakevdp 2020-09-18
api_benchmark results:
master branch:
```
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
jit_trivial_dispatch              120578 ns       120539 ns         5381
jit_trivial                       136459 ns       136437 ns         5146
jit_simple_dispatch               495457 ns       494981 ns         1461
jit_simple                        510835 ns       510391 ns         1185
jit_simple_many_args_dispatch    2863614 ns      2863011 ns          246
jit_simple_many_args             2872962 ns      2872501 ns          234
jit_dispatch_without_transfer    6348334 ns      6347968 ns           91
jit_dispatch_with_transfer       6126572 ns      6126012 ns          117
```
with this change:
```
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
jit_trivial_dispatch              106872 ns       106829 ns         6227
jit_trivial                       127626 ns       127625 ns         5493
jit_simple_dispatch               499879 ns       499377 ns         1417
jit_simple                        512144 ns       512095 ns         1342
jit_simple_many_args_dispatch    3088399 ns      3087974 ns          227
jit_simple_many_args             3043566 ns      3042474 ns          226
jit_dispatch_without_transfer    6320484 ns      6320214 ns          107
jit_dispatch_with_transfer       6268184 ns      6266837 ns          116

```